+++ .env.example
@@ -22,6 +22,6 @@ GOOGLE_CLIENT_ID="532789583215-87s4ctpgglpk24bn76gani097m2od2ae.apps.googleuserc
 GOOGLE_CLIENT_SECRET="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"
 
 #openai
-OPENAI_API_KEY="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"
+OPENAI_API_KEY=""
