+++ src/@server/config/env-schema.ts
@@ -6,6 +6,7 @@ const CONFIG_ENV_SCHEMA = z.object({
   NEXTAUTH_SECRET: z.string(),
   GOOGLE_CLIENT_ID: z.string(),
   GOOGLE_CLIENT_SECRET: z.string(),
+  OPENAI_API_KEY: z.string(),
 })
 
 export const env = CONFIG_ENV_SCHEMA.parse(process.env)
