+++ src/@server/modules/Game/infra/prisma/repositories/GamesRepository.ts
@@ -3,7 +3,7 @@ import 'reflect-metadata';
 import { Prisma } from '@prisma/client';
 import { DefaultArgs } from '@prisma/client/runtime/library';
 
-import AppDataSource from '@/@server/shared/infra/http/db/data-source';
+import AppDataSource from '@/@server/config/data-source';
 import IFindAllProvidersDTO from '../../../dtos/IFindAllProvidersDTO';
 import IGamesRepository from '../../../repositories/IGamesRepository';
 import Game from '../entities/Game';
