+++ src/@server/modules/Questions/repositories/IOpenAIApiRepository.ts
@@ -1,6 +1,5 @@
 import ICreateChatCompletionDTO from "../dtos/ICreateChatCompletionDTO";
-import Question from "../infra/prisma/entities/Question";
 
 export default interface IOpenAIApiRepository {
-  createChatCompletion(data: ICreateChatCompletionDTO): Promise<Question[]>;
+  createChatCompletion(data: ICreateChatCompletionDTO): Promise<any>;
 }
