--- src/redux/actions/questions/get-all-questions.ts
@@ -1,13 +1,0 @@
-import { api } from "@/lib/api";
-
-const getAllQuestions= async (): Promise<QuestionData[]> => {
-  const response = await api.get("/api/questions/");
-
-  return response.data;
-};
-
-const getAllQuestionsAction = {
-  getAllQuestions,
-}
-
-export default getAllQuestionsAction;
