+++ src/redux/store.ts
@@ -5,7 +5,6 @@ import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
 import { FLUSH, PAUSE, PERSIST, PURGE, REGISTER, REHYDRATE, persistReducer } from 'redux-persist';
 
 import gamesSlice from "@/redux/slices/games-slice";
-import questionsSlice from "@/redux/slices/questions-slice";
 import topicsSlice from "@/redux/slices/topics-slice";
 
 const rootPersistConfig = {
@@ -15,7 +14,6 @@ const rootPersistConfig = {
 };
 
 const rootReducer = combineReducers({
-  questions: questionsSlice,
   topics: topicsSlice,
   games: gamesSlice,
 });
